+++ src/modules/CVs/API/CVs.ts
@@ -3,21 +3,11 @@ import {
     generateDeleteMutation,
     generateCreateMutation,
     generateUpdateMutation,
-} from "../helpers/queries";
+} from "@/Api/helpers/queries";
 
-const query = "project";
+const query = "cv";
 
-const fields = [
-    "id",
-    "created_at",
-    "name",
-    "internal_name",
-    "description",
-    "domain",
-    "start_date",
-    "end_date",
-    "team_size",
-];
+const fields = ["id", "created_at", "name", "education", "description", "user"];
 const createArgs = [
     "name",
     "internal_name",
