+++ tsconfig.json
@@ -17,7 +17,10 @@
         "noUnusedLocals": true,
         "noUnusedParameters": true,
         "noFallthroughCasesInSwitch": true,
-        "baseUrl": "."
+        "baseUrl": ".",
+        "paths": {
+            "@/*": ["./src/*"]
+        }
     },
     "include": ["src/**/*", "vite.config.ts", "*.d.ts"],
     "exclude": ["node_modules", "dist"],
